requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
pyquery
pyyaml
requests
